% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_usd_to_eur.R
\name{convert_usd_to_eur}
\alias{convert_usd_to_eur}
\title{Exchange rate for a certain amount in the latest day}
\usage{
convert_usd_to_eur(usd)
}
\arguments{
\item{usd}{the amount in dollars}
}
\value{
number (the amount in EUR)
}
\description{
Exchange rate for a certain amount in the latest day
}
\examples{
convert_usd_eur(54)
}
