% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eurusds.R
\name{eurusds}
\alias{eurusds}
\title{Exchange rate for a certain day}
\usage{
eurusds(date_from, date_to)
}
\arguments{
\item{date_from}{first date we are interested on year-month-date format}

\item{date_to}{last date we are interested on year-month-date format}
}
\value{
data.table object
}
\description{
Exchange rate for a certain day
}
