% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uneur.R
\name{uneur}
\alias{uneur}
\title{strips numbers from its Euro currency}
\usage{
uneur(x)
}
\arguments{
\item{x}{number with euro sign}
}
\value{
only number
}
\description{
strips numbers from its Euro currency
}
\examples{
uneur("€100,000")
uneur("€46,546.99")
}
